added data and styling
