+++ src/theme.tsx
@@ -16,6 +16,26 @@ let theme = createTheme({
       default: "#fff",
     },
   },
+  typography: {
+    h1: {
+      fontSize: "3rem",
+    },
+    h2: {
+      fontSize: "2.5rem",
+    },
+    h3: {
+      fontSize: "2rem",
+    },
+    h4: {
+      fontSize: "1.5rem",
+    },
+    h5: {
+      fontSize: "1.2rem",
+    },
+    h6: {
+      fontSize: "1.1rem",
+    },
+  },
 });
 
 theme = responsiveFontSizes(theme);
